requirements.txt
Cerberus==1.2
Flask==1.0.2
Flask-Cors==3.0.7
Jinja2==2.10
py-bcrypt==0.4
pymongo==3.7.2
python-dotenv==0.10.1
Werkzeug==0.14.1
Celery==4.2.0
redis==3.1.0
Pillow==5.4.1
numpy==1.16.1
python-dotenv==0.10.1
